+++ modules/algorithms/src/kmeans.cpp
@@ -77,7 +77,7 @@ void Kmeans::learn(cv::Mat_<float>& input, ClusteringParams* parameters){
   for(int i = 0; i < labels.rows; ++i){
     clusters[labels.at<int>(i, 0)].push_back(i);
   }
-  for(int i = 0; i < clusters.size(); ++i){
+  for(int i = 0; i < static_cast<int>(clusters.size()); ++i){
     auto cluster = clusters[i];
     clusters_.push_back(cluster);
   }
@@ -141,7 +141,7 @@ void Kmeans::clear(){
 void Kmeans::setupLabelMatFromInitialization(cv::Mat& labels){
   if(clusters_.empty()) return;
   labels = cv::Mat_<int>::zeros(samples_.rows, samples_.cols);
-  for(int c = 0; c < clusters_.size(); ++c){
+  for(int c = 0; c < static_cast<int>(clusters_.size()); ++c){
     for(int s = 0; s < clusters_[c].size(); ++s){
       labels.at<int>(clusters_[c][s], 0) = c;
     }
